Match recordings using longest common subsequence
This fixes mismatched recordings that occur especially with
multi-day singings where only one day was recorded (e.g. 2013 AL
State only has a recording of Sunday, but recordings were being
assigned to both Saturday and Sunday singings.

This requires recordings to be listed in order, but not all
recordings are in order (e.g. 2013 United lists Sunday before
Saturday).  To get around this, each day in a multi-day singing
is matched separately.