Avoid commiting after each leader insert
Most of the parse time is in db commits.  Since leaders are added
as the minutes are parsed, we keep track of leader names in a
dict then commit at the end of parsing.

To speed up inserts a little more, songs, aliases, and invalid
leaders are pulled into a dict from the db once.